modules/aadauth/requirements.txt
adal
flask
requests
pyyaml